requirements.txt
streamlit>=1.32.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
pandas>=2.0.0
plotly>=5.18.0
extra-streamlit-components>=0.1.71
PyJWT>=2.8.0
